style: move to module directories
